removed error_handling.hpp, moving functions into cufile.hpp
