+++ cudax/include/cuda/experimental/__cufile/buffer_handle.hpp
@@ -21,7 +21,7 @@
 
 #include <cuda/std/span>
 
-#include <cuda/experimental/__cufile/detail/error_handling.hpp>
+#include <cuda/experimental/__cufile/cufile.hpp>
 
 #include <functional>
