+++ cudax/include/cuda/experimental/__cufile/cufile.hpp
@@ -12,23 +12,87 @@
 // cufile.hpp — Modern C++ bindings for NVIDIA cuFILE (GPU Direct Storage)
 // Provides a clean interface that directly maps to the cuFILE C API.
 
-// ================================================================================================
-// Core Components
-// ================================================================================================
+#include <cuda/std/detail/__config>
 
-#include <cuda/experimental/__cufile/detail/driver.hpp> // Driver management and configuration
-#include <cuda/experimental/__cufile/detail/file_handle.hpp> // File operations
-#include <cuda/experimental/__cufile/detail/utils.hpp> // Utility functions
+#if defined(_CCCL_IMPLICIT_SYSTEM_HEADER_GCC)
+#  pragma GCC system_header
+#elif defined(_CCCL_IMPLICIT_SYSTEM_HEADER_CLANG)
+#  pragma clang system_header
+#elif defined(_CCCL_IMPLICIT_SYSTEM_HEADER_MSVC)
+#  pragma system_header
+#endif // no system header
+
+#include <cuda/experimental/__cufile/detail/enums.hpp>
+
+#include <stdexcept>
+#include <string>
+
+#include <cufile.h>
 
 // CUDA Experimental cuFILE Library namespace
 namespace cuda::experimental::cufile
 {
+// Forward declare driver API used by helpers below
+void driver_open();
+void driver_close();
+long driver_use_count();
+int get_version();
 
 // ================================================================================================
 // Error Handling
 // ================================================================================================
 
-using cufile_exception = detail::cufile_exception;
+//! Unified cuFile exception class
+class cufile_exception : public ::std::runtime_error
+{
+private:
+  CUfileError_t error_;
+
+public:
+  explicit cufile_exception(CUfileError_t error)
+      : ::std::runtime_error(format_error_message(error))
+      , error_(error)
+  {}
+
+  explicit cufile_exception(const ::std::string& message)
+      : ::std::runtime_error(message)
+      , error_{to_c_enum(cu_file_error::success), CUDA_SUCCESS}
+  {}
+
+  CUfileError_t error() const noexcept
+  {
+    return error_;
+  }
+
+private:
+  static ::std::string format_error_message(CUfileError_t error)
+  {
+    return ::std::string("cuFile error: ") + ::std::to_string(error.err) + " (CUDA: " + ::std::to_string(error.cu_err)
+         + ")";
+  }
+};
+
+//! Check cuFile operation result and throw on error
+inline void check_cufile_result(CUfileError_t error, const ::std::string& operation = "")
+{
+  if (error.err != to_c_enum(cu_file_error::success))
+  {
+    ::std::string message = operation.empty() ? "" : operation + ": ";
+    throw cufile_exception(error);
+  }
+}
+
+//! Check cuFile operation result and throw on error (for ssize_t returns)
+inline ssize_t check_cufile_result(ssize_t result, const ::std::string& operation = "")
+{
+  if (result < 0)
+  {
+    CUfileError_t error   = {to_c_enum_from_result(result), CUDA_SUCCESS};
+    ::std::string message = operation.empty() ? "" : operation + ": ";
+    throw cufile_exception(error);
+  }
+  return result;
+}
 
 //! Initialize the cuFILE library
 inline void initialize()
@@ -37,14 +101,9 @@ inline void initialize()
 }
 
 //! Shutdown the cuFILE library
-inline void shutdown() noexcept
+inline void shutdown()
 {
-  try
-  {
-    driver_close();
-  }
-  catch (...)
-  {}
+  driver_close();
 }
 
 //! Check if the cuFILE library is initialized
@@ -59,4 +118,28 @@ inline int get_cufile_version() noexcept
   return get_version();
 }
 
+namespace detail
+{
+inline void check_cufile_result(CUfileError_t error, const ::std::string& operation = "")
+{
+  ::cuda::experimental::cufile::check_cufile_result(error, operation);
+}
+
+inline ssize_t check_cufile_result(ssize_t result, const ::std::string& operation = "")
+{
+  return ::cuda::experimental::cufile::check_cufile_result(result, operation);
+}
+} // namespace detail
+
 } // namespace cuda::experimental::cufile
+
+// ================================================================================================
+// Core Components
+// ================================================================================================
+
+#include <cuda/experimental/__cufile/batch_handle.hpp>
+#include <cuda/experimental/__cufile/buffer_handle.hpp>
+#include <cuda/experimental/__cufile/driver.hpp>
+#include <cuda/experimental/__cufile/file_handle.hpp>
+#include <cuda/experimental/__cufile/stream_handle.hpp>
+#include <cuda/experimental/__cufile/utils.hpp>
